Move fields and fallbacks into their own modules and update docs to reflect this.
